requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
drf-yasg==1.21.11
environs==14.3.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
marshmallow==4.0.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
